simplify and improve plot
This makes the webpage look a lot nicer.  I think the webpage builder,
pkgdown, must be doing a resize, and that ruins the colour-fill scheme
(which is using par("usr"), which I guess is not updated upon a resize).

I ought to figure out why pkgdown does that, because it also makes the
oce webpages look ugly.  Another day, though.