use R-CMD-check as per oce
It sure is annoying to have to keep updating these action files, making
guess after guess based on search after search.  Is any of this actually
documented?  And, after years of this, why haven't things settled down
more?  It woiuld help if old versions of things ought to give some kind
of a hint that they need updating.