fix hour-old error in knockdown
I had forgotten to remove the "1/g" factor that I had for drag().  I had
that before because I had put gravity there instead of in buoyancy(). It
all works out in the wash -- the final results are as before, but now
the formulae are in Newtons, which makes more sense.  (This is a
force-balance problem, so it does not matter in what unit I express
forces so long as I am consistent -- that is why the results are not
changed from 2 hours ago.)